+++ pallets/living-assets-ownership/Cargo.toml
@@ -12,6 +12,7 @@ edition = "2021"
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+hex-literal = { workspace = true, optional = true}
 parity-scale-codec = { workspace = true, features = ["derive"] }
 scale-info = { workspace = true, features = ["derive"] }
 
@@ -24,6 +25,7 @@ sp-core = { workspace = true }
 
 [dev-dependencies]
 serde = { workspace = true }
+hex = { version = "0.4.3" }
 
 # Substrate
 sp-core = { workspace = true }
